Remove local LLM support files
